git-svn-id: http://svn.olex2.org/olex2-gui/branches/1.5@6392 bceb41a0-0b25-4981-bef4-7ef80ce4c728

+++ util/pyUtil/htmlMaker.py
@@ -623,7 +623,7 @@ def restraint_builder(cmd):
   controls.append('$spy.MakeHoverButton("button_small-go@%s","%s")' %(name, onclick))
 
   html = ""
-  colw = 100/len(controls)
+  colw = int(100/len(controls))
   for i, td in enumerate(controls):
     if (i+1) == len(controls):
       align = "right"
